Rename org.eclipse.leshan.json package in org.eclipse.leshan.core.json
