+++ leshan-core/src/test/java/org/eclipse/leshan/json/JsonSerializerTest.java
@@ -18,6 +18,10 @@ package org.eclipse.leshan.json;
 
 import java.util.ArrayList;
 
+import org.eclipse.leshan.core.json.JsonArrayEntry;
+import org.eclipse.leshan.core.json.JsonRootObject;
+import org.eclipse.leshan.core.json.LwM2mJson;
+import org.eclipse.leshan.core.json.LwM2mJsonException;
 import org.junit.Assert;
 import org.junit.Test;
 import org.slf4j.Logger;
